bugfix (for cygwin): CPPUNIT_ASSERT_EQUAL requires the types of the values to be compared to be exactly the same

+++ src/tests/SmartPtrTestTemplate.cpp
@@ -154,18 +154,18 @@ namespace wns { namespace tests {
 	void SmartPtrTest::refCountable()
 	{
 		A a;
-		CPPUNIT_ASSERT_EQUAL( 0, a.getRefCount() );
+		CPPUNIT_ASSERT_EQUAL( static_cast<int32_t>(0), a.getRefCount() );
 	}
 
 	void SmartPtrTest::STLContainer()
 	{
 		SmartPtr<A> spt(new A);
 		std::map<SmartPtr<A>, int> container;
-		CPPUNIT_ASSERT_EQUAL( 1, spt.getPtr()->getRefCount() );
+		CPPUNIT_ASSERT_EQUAL( static_cast<int32_t>(1), spt.getPtr()->getRefCount() );
 		container[spt] = 5;
-		CPPUNIT_ASSERT_EQUAL( 2, spt.getPtr()->getRefCount() );
+		CPPUNIT_ASSERT_EQUAL( static_cast<int32_t>(2), spt.getPtr()->getRefCount() );
 		container.clear();
-		CPPUNIT_ASSERT_EQUAL( 1, spt.getPtr()->getRefCount() );
+		CPPUNIT_ASSERT_EQUAL( static_cast<int32_t>(1), spt.getPtr()->getRefCount() );
 	}
 
 	void SmartPtrTest::constructor()
@@ -177,38 +177,38 @@ namespace wns { namespace tests {
 
 		{
 			SmartPtr<A> a(new A);
-			CPPUNIT_ASSERT_EQUAL( 1, a.getRefCount() );
+			CPPUNIT_ASSERT_EQUAL( static_cast<int32_t>(1), a.getRefCount() );
 		}
 		{
 			SmartPtr<A> a(new A);
-			CPPUNIT_ASSERT_EQUAL( 1, a.getRefCount() );
+			CPPUNIT_ASSERT_EQUAL( static_cast<int32_t>(1), a.getRefCount() );
 			SmartPtr<A> a2(a);
-			CPPUNIT_ASSERT_EQUAL( 2, a.getRefCount() );
-			CPPUNIT_ASSERT_EQUAL( 2, a2.getRefCount() );
+			CPPUNIT_ASSERT_EQUAL( static_cast<int32_t>(2), a.getRefCount() );
+			CPPUNIT_ASSERT_EQUAL( static_cast<int32_t>(2), a2.getRefCount() );
 		}
 
 		{
 			A* aPtr = new A;
 			SmartPtr<A> a(aPtr);
-			CPPUNIT_ASSERT_EQUAL( 1, a.getRefCount() );
+			CPPUNIT_ASSERT_EQUAL( static_cast<int32_t>(1), a.getRefCount() );
 		}
 	}
 
 	void SmartPtrTest::operators()
 	{
 		SmartPtr<A> a(new A);
-		CPPUNIT_ASSERT_EQUAL( 42, a->foo() );
+		CPPUNIT_ASSERT_EQUAL( static_cast<int32_t>(42), a->foo() );
 		SmartPtr<A> a2 = a;
-		CPPUNIT_ASSERT_EQUAL( 2, a2.getRefCount() );
-		CPPUNIT_ASSERT_EQUAL( 2, a.getRefCount() );
+		CPPUNIT_ASSERT_EQUAL( static_cast<int32_t>(2), a2.getRefCount() );
+		CPPUNIT_ASSERT_EQUAL( static_cast<int32_t>(2), a.getRefCount() );
 		CPPUNIT_ASSERT( a==a2 );
 		a2 = SmartPtr<A>(new A);
 		CPPUNIT_ASSERT( a!=a2 );
-		CPPUNIT_ASSERT_EQUAL( 1, a2.getRefCount() );
-		CPPUNIT_ASSERT_EQUAL( 1, a.getRefCount() );
+		CPPUNIT_ASSERT_EQUAL( static_cast<int32_t>(1), a2.getRefCount() );
+		CPPUNIT_ASSERT_EQUAL( static_cast<int32_t>(1), a.getRefCount() );
 		A copyOfA = *a;
 		// make sure a copy of the object has refcount 0;
-		CPPUNIT_ASSERT_EQUAL( 0, copyOfA.getRefCount() );
+		CPPUNIT_ASSERT_EQUAL( static_cast<int32_t>(0), copyOfA.getRefCount() );
 
 		CPPUNIT_ASSERT( a );
 		SmartPtr<A> a3;
@@ -222,15 +222,15 @@ namespace wns { namespace tests {
 		CPPUNIT_ASSERT( !destructorHasBeenCalled );
 		{
 			SmartPtr<B> b(bPtr);
-			CPPUNIT_ASSERT_EQUAL( 1, b.getRefCount() );
+			CPPUNIT_ASSERT_EQUAL( static_cast<int32_t>(1), b.getRefCount() );
 			CPPUNIT_ASSERT( !destructorHasBeenCalled );
 			{
 				SmartPtr<B> b2 = b;
-				CPPUNIT_ASSERT_EQUAL( 2, b2.getRefCount() );
-				CPPUNIT_ASSERT_EQUAL( 2, b.getRefCount() );
+				CPPUNIT_ASSERT_EQUAL( static_cast<int32_t>(2), b2.getRefCount() );
+				CPPUNIT_ASSERT_EQUAL( static_cast<int32_t>(2), b.getRefCount() );
 				CPPUNIT_ASSERT( !destructorHasBeenCalled );
 			}
-			CPPUNIT_ASSERT_EQUAL( 1, b.getRefCount() );
+			CPPUNIT_ASSERT_EQUAL( static_cast<int32_t>(1), b.getRefCount() );
 			CPPUNIT_ASSERT( !destructorHasBeenCalled );
 		}
 		CPPUNIT_ASSERT( destructorHasBeenCalled );
@@ -239,7 +239,7 @@ namespace wns { namespace tests {
 		D* dPtr = new D(this);
 		{
 			SmartPtr<D> d(dPtr);
-			CPPUNIT_ASSERT_EQUAL( 1, d.getRefCount() );
+			CPPUNIT_ASSERT_EQUAL( static_cast<int32_t>(1), d.getRefCount() );
 		}
 		CPPUNIT_ASSERT( destructorHasBeenCalled );
 	}
@@ -247,13 +247,13 @@ namespace wns { namespace tests {
 	void SmartPtrTest::upCast()
 	{
 		SmartPtr<A> a(new C);
-		CPPUNIT_ASSERT_EQUAL( 1, a.getRefCount() );
+		CPPUNIT_ASSERT_EQUAL( static_cast<int32_t>(1), a.getRefCount() );
 		SmartPtr<C> c(new C);
 		CPPUNIT_ASSERT( a!=c );
-		CPPUNIT_ASSERT_EQUAL( 1, c.getRefCount() );
+		CPPUNIT_ASSERT_EQUAL( static_cast<int32_t>(1), c.getRefCount() );
 		a = c;
-		CPPUNIT_ASSERT_EQUAL( 2, a.getRefCount() );
-		CPPUNIT_ASSERT_EQUAL( 2, c.getRefCount() );
+		CPPUNIT_ASSERT_EQUAL( static_cast<int32_t>(2), a.getRefCount() );
+		CPPUNIT_ASSERT_EQUAL( static_cast<int32_t>(2), c.getRefCount() );
 		CPPUNIT_ASSERT( a==c );
 	}
